renamed concrete_core to concrete and copied swole's project orchestrator separation work
