+++ pyproject.toml
@@ -15,7 +15,7 @@ dev-dependencies = [
 ]
 
 [tool.uv.workspace]
-members = ["src/concrete-core", "src/concrete-async", "src/concrete-db"] # Can't do src/* because of eggs and misc auto-generated files
+members = ["src/concrete-core", "src/concrete-async", "src/concrete-db", "webapp/api"] # Can't do src/* because of eggs and misc auto-generated files
 
 [tool.uv.sources]
 concrete-core = { workspace = true }
