+++ src/concrete-core/concrete/__init__.py
@@ -1,8 +1,8 @@
 from dotenv import load_dotenv
 
-from . import operators, orchestrator
+from . import operators, orchestrators
 
 # Always runs even when importing submodules
 # https://stackoverflow.com/a/27144933
 load_dotenv(override=True)
-__all__ = ["operators", "orchestrator"]
+__all__ = ["operators", "orchestrators"]
