+++ src/concrete-core/concrete/models/messages.py
@@ -61,11 +61,6 @@ class TextMessage(Message):
     text: str = Field()
 
 
-class Param(Message):
-    name: str = Field(description='Name of the parameter')
-    value: str = Field(description="Value of the parameter")
-
-
 class Tool(Message):
     tool_name: str = Field(description='Name of the tool')
     tool_method: str = Field(description='Command to call the tool')
