+++ src/concrete-core/concrete/orchestrators/__init__.py
@@ -1,0 +1,8 @@
+from .software_orchestrator import SoftwareOrchestrator
+
+
+class Orchestrator:
+    pass
+
+
+__all__ = ["SoftwareOrchestrator", "Orchestrator"]
