+++ src/concrete-core/concrete/prompts.py
@@ -4,13 +4,13 @@ Generate a website using Flask which has an input form at the top.
 When the input form is submitted, it should call `invoke_concrete` with the form input.
 Include the following function `invoke_concrete` with the server code:
 ```python
-from concrete import orchestrator
+from concrete import orchestrators
 
 def invoke_concrete(input_str: str):
     '''
     Returns a valid html element
     '''
-    so = orchestrator.SoftwareOrchestrator()
+    so = orchestrators.SoftwareOrchestrator()
     element = so.agents['dev'].implement_html_element(input_str)
     return "\n".join(element.strip().split('\n')[1:-1])
 ```
