+++ tests/test_dependencies.py
@@ -2,7 +2,7 @@ import asyncio
 import sys
 
 import pytest
-from concrete_core.orchestrator import SoftwareOrchestrator
+from concrete.orchestrators import SoftwareOrchestrator
 
 
 class TestDependencies:
