+++ tests/test_models.py
@@ -1,8 +1,8 @@
 import unittest
 from uuid import uuid4
 
-from concrete_core.models.messages import TextMessage
-from concrete_core.operators import Developer, Executive
+from concrete.models.messages import TextMessage
+from concrete.operators import Developer, Executive
 from concrete_db.orm.models import Message as SQLModelMessage
 from concrete_db.orm.models import Operator as SQLModelOperator
 from concrete_db.orm.models import Project
