+++ webapp/api/hello.py
@@ -1,0 +1,6 @@
+def main():
+    print("Hello from api!")
+
+
+if __name__ == "__main__":
+    main()
