+++ webapp/api/server.py
@@ -4,11 +4,9 @@ from typing import Annotated
 from uuid import UUID
 
 import dotenv
-from concrete_core.webutils import AuthMiddleware
+from concrete.webutils import AuthMiddleware
 from concrete_db import crud
-from concrete_db.orm import (
-    SessionLocal,  # TODO Stop using SessionLocal in favor of Session context manager
-)
+from concrete_db.orm import Session
 from concrete_db.orm.models import (
     Client,
     ClientCreate,
@@ -21,7 +19,6 @@ from concrete_db.orm.models import (
     OrchestratorUpdate,
 )
 from fastapi import Depends, FastAPI, HTTPException
-from sqlmodel import Session
 from starlette.middleware import Middleware
 from starlette.middleware.sessions import SessionMiddleware
 
@@ -50,17 +47,6 @@ If the db already exists and the sql models are the same, then behavior is as ex
 If the db already exists but the sql models differ, then migrations will need to be run for DB interaction
 to function as expected.
 """
-
-
-def get_db():
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
-
-
-DbDep = Annotated[Session, Depends(get_db)]
 
 
 def get_common_read_params(skip: int = 0, limit: int = 100) -> CommonReadParameters:
@@ -92,149 +78,173 @@ def ping():
 
 # ===CRUD operations for Orchestrators=== #
 @app.post("/orchestrators/", response_model=Orchestrator)
-def create_orchestrator(orchestrator: OrchestratorCreate, db: DbDep) -> Orchestrator:
-    return crud.create_orchestrator(db, orchestrator)
+def create_orchestrator(orchestrator: OrchestratorCreate) -> Orchestrator:
+    with Session() as db:
+        return crud.create_orchestrator(db, orchestrator)
 
 
 @app.get("/orchestrators/")
-def get_orchestrators(common_read_params: CommonReadDep, db: DbDep) -> Sequence[Orchestrator]:
-    return crud.get_orchestrators(
-        db,
-        skip=common_read_params.skip,
-        limit=common_read_params.limit,
-    )
+def get_orchestrators(common_read_params: CommonReadDep) -> Sequence[Orchestrator]:
+    with Session() as db:
+        return crud.get_orchestrators(
+            db,
+            skip=common_read_params.skip,
+            limit=common_read_params.limit,
+        )
 
 
 @app.get("/orchestrators/{orchestrator_id}")
-def get_orchestrator(orchestrator_id: UUID, db: DbDep) -> Orchestrator:
-    orchestrator = crud.get_orchestrator(db, orchestrator_id)
-    if orchestrator is None:
-        raise orchestrator_not_found(orchestrator_id)
-    return orchestrator
+def get_orchestrator(orchestrator_id: UUID) -> Orchestrator:
+    with Session() as db:
+        orchestrator = crud.get_orchestrator(db, orchestrator_id)
+        if orchestrator is None:
+            raise orchestrator_not_found(orchestrator_id)
+        return orchestrator
 
 
 @app.put("/orchestrators/{orchestrator_id}")
-def update_orchestrator(orchestrator_id: UUID, orchestrator: OrchestratorUpdate, db: DbDep) -> Orchestrator:
-    db_orc = crud.update_orchestrator(db, orchestrator_id, orchestrator)
-    if db_orc is None:
-        raise orchestrator_not_found(orchestrator_id)
-    return db_orc
+def update_orchestrator(orchestrator_id: UUID, orchestrator: OrchestratorUpdate) -> Orchestrator:
+    with Session() as db:
+        db_orc = crud.update_orchestrator(db, orchestrator_id, orchestrator)
+        if db_orc is None:
+            raise orchestrator_not_found(orchestrator_id)
+        return db_orc
 
 
 @app.delete("/orchestrators/{orchestrator_id}")
-def delete_orchestrator(orchestrator_id: UUID, db: DbDep) -> Orchestrator:
-    orchestrator = crud.delete_orchestrator(db, orchestrator_id)
-    if orchestrator is None:
-        raise orchestrator_not_found(orchestrator_id)
-    return orchestrator
+def delete_orchestrator(orchestrator_id: UUID) -> Orchestrator:
+    with Session() as db:
+        orchestrator = crud.delete_orchestrator(db, orchestrator_id)
+        if orchestrator is None:
+            raise orchestrator_not_found(orchestrator_id)
+        return orchestrator
 
 
 # ===CRUD operations for Operators=== #
 @app.post("/operators/")
-def create_operator(operator: OperatorCreate, db: DbDep) -> Operator:
-    orchestrator = crud.get_orchestrator(db, operator.orchestrator_id)
-    if orchestrator is None:
-        raise orchestrator_not_found(operator.orchestrator_id)
-    return crud.create_operator(db, operator)
+def create_operator(operator: OperatorCreate) -> Operator:
+    with Session() as db:
+        orchestrator = crud.get_orchestrator(db, operator.orchestrator_id)
+        if orchestrator is None:
+            raise orchestrator_not_found(operator.orchestrator_id)
+        return crud.create_operator(db, operator)
 
 
 @app.get("/operators/")
-def read_operators(common_read_params: CommonReadDep, db: DbDep) -> Sequence[Operator]:
-    return crud.get_operators(
-        db,
-        skip=common_read_params.skip,
-        limit=common_read_params.limit,
-    )
+def read_operators(common_read_params: CommonReadDep) -> Sequence[Operator]:
+    with Session() as db:
+        return crud.get_operators(
+            db,
+            skip=common_read_params.skip,
+            limit=common_read_params.limit,
+        )
 
 
 @app.get("/orchestrators/{orchestrator_id}/operators/")
 def read_orchestrator_operators(
     orchestrator_id: UUID,
     common_read_params: CommonReadDep,
-    db: DbDep,
 ) -> Sequence[Operator]:
-    return crud.get_operators(
-        db,
-        orchestrator_id,
-        common_read_params.skip,
-        common_read_params.limit,
-    )
+    with Session() as db:
+        return crud.get_operators(
+            db,
+            orchestrator_id,
+            common_read_params.skip,
+            common_read_params.limit,
+        )
 
 
 @app.get("/orchestrators/{orchestrator_id}/operators/{operator_id}")
-def read_operator(orchestrator_id: UUID, operator_id: UUID, db: DbDep) -> Operator:
-    operator = crud.get_operator(db, operator_id, orchestrator_id)
-    if operator is None:
-        raise operator_not_found(operator_id)
-    return operator
+def read_operator(orchestrator_id: UUID, operator_id: UUID) -> Operator:
+    with Session() as db:
+        operator = crud.get_operator(db, operator_id, orchestrator_id)
+        if operator is None:
+            raise operator_not_found(operator_id)
+        return operator
 
 
 @app.put("/orchestrators/{orchestrator_id}/operators/{operator_id}")
-def update_operator(orchestrator_id: UUID, operator_id: UUID, operator: OperatorUpdate, db: DbDep) -> Operator:
-    db_op = crud.update_operator(db, operator_id, orchestrator_id, operator)
-    if db_op is None:
-        raise operator_not_found(operator_id)
-    return db_op
+def update_operator(orchestrator_id: UUID, operator_id: UUID, operator: OperatorUpdate) -> Operator:
+    with Session() as db:
+        db_operator = crud.update_operator(db, operator_id, orchestrator_id, operator)
+        if db_operator is None:
+            raise operator_not_found(operator_id)
+        return db_operator
 
 
 @app.delete("/orchestrators/{orchestrator_id}/operators/{operator_id}")
-def delete_operator(orchestrator_id: UUID, operator_id: UUID, db: DbDep) -> Operator:
-    operator = crud.delete_operator(db, operator_id, orchestrator_id)
-    if operator is None:
-        raise operator_not_found(operator_id)
-    return operator
+def delete_operator(orchestrator_id: UUID, operator_id: UUID) -> Operator:
+    with Session() as db:
+        operator = crud.delete_operator(db, operator_id, orchestrator_id)
+        if operator is None:
+            raise operator_not_found(operator_id)
+        return operator
 
 
 # ===CRUD operations for Clients=== #
 @app.post("/clients/")
-def create_client(client: ClientCreate, db: DbDep) -> Client:
-    operator = crud.get_operator(db, client.operator_id, client.orchestrator_id)
-    if operator is None:
-        raise operator_not_found(client.operator_id)
-    return crud.create_client(db, client)
+def create_client(client: ClientCreate) -> Client:
+    with Session() as db:
+        operator = crud.get_operator(db, client.operator_id, client.orchestrator_id)
+        if operator is None:
+            raise operator_not_found(client.operator_id)
+        return crud.create_client(db, client)
 
 
 @app.get("/clients/")
-def read_clients(common_read_params: CommonReadDep, db: DbDep) -> Sequence[Client]:
-    return crud.get_clients(
-        db,
-        skip=common_read_params.skip,
-        limit=common_read_params.limit,
-    )
+def read_clients(common_read_params: CommonReadDep) -> Sequence[Client]:
+    with Session() as db:
+        return crud.get_clients(
+            db,
+            skip=common_read_params.skip,
+            limit=common_read_params.limit,
+        )
 
 
 @app.get("/orchestrators/{orchestrator_id}/operators/{operator_id}/clients/")
 def read_operator_clients(
-    orchestrator_id: UUID, operator_id: UUID, common_read_params: CommonReadDep, db: DbDep
+    orchestrator_id: UUID,
+    operator_id: UUID,
+    common_read_params: CommonReadDep,
 ) -> Sequence[Client]:
-    return crud.get_clients(
-        db,
-        orchestrator_id=orchestrator_id,
-        operator_id=operator_id,
-        skip=common_read_params.skip,
-        limit=common_read_params.limit,
-    )
+    with Session() as db:
+        return crud.get_clients(
+            db,
+            orchestrator_id=orchestrator_id,
+            operator_id=operator_id,
+            skip=common_read_params.skip,
+            limit=common_read_params.limit,
+        )
 
 
 @app.get("/orchestrators/{orchestrator_id}/operators/{operator_id}/clients/{client_id}")
-def read_client(orchestrator_id: UUID, operator_id: UUID, client_id: UUID, db: DbDep) -> Client:
-    client = crud.get_client(db, client_id, operator_id, orchestrator_id)
-    if client is None:
-        raise client_not_found(client_id)
-    return client
+def read_client(orchestrator_id: UUID, operator_id: UUID, client_id: UUID) -> Client:
+    with Session() as db:
+        client = crud.get_client(db, client_id, operator_id, orchestrator_id)
+        if client is None:
+            raise client_not_found(client_id)
+        return client
 
 
 @app.put("/orchestrator/{orchestrator_id}/operators/{operator_id}/clients/{client_id}")
-def update_client(orchestrator_id: UUID, operator_id: UUID, client_id: UUID, client: ClientUpdate, db: DbDep) -> Client:
-    db_client = crud.update_client(db, client_id, operator_id, orchestrator_id, client)
-    if db_client is None:
-        raise client_not_found(client_id)
-    return db_client
+def update_client(orchestrator_id: UUID, operator_id: UUID, client_id: UUID, client: ClientUpdate) -> Client:
+    with Session() as db:
+        db_client = crud.update_client(db, client_id, operator_id, orchestrator_id, client)
+        if db_client is None:
+            raise client_not_found(client_id)
+        return db_client
 
 
 @app.delete("/orchestrator/{orchestrator_id}/operators/{operator_id}/clients/{client_id}")
-def delete_client(orchestrator_id: UUID, operator_id: UUID, client_id: UUID, db: DbDep) -> Client:
-    client = crud.delete_client(db, client_id, operator_id, orchestrator_id)
-    if client is None:
-        raise client_not_found(client_id)
-    return client
+def delete_client(orchestrator_id: UUID, operator_id: UUID, client_id: UUID) -> Client:
+    with Session() as db:
+        client = crud.delete_client(db, client_id, operator_id, orchestrator_id)
+        if client is None:
+            raise client_not_found(client_id)
+        return client
+
+
+# ===Project and Operator Building=== #
+
+# @app.post("/build/project")
+# def init_project(name: str) -> Projec
